% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Drugs_SigsR}
\alias{Drugs_SigsR}
\title{Drugs_SigsR}
\usage{
Drugs_SigsR(datasetInput_Dataset = NULL, refDrug = NULL, n_bins = 33,
  datasetInput_Dis = NULL, ...)
}
\arguments{
\item{datasetInput_Dataset}{Specify a dataset of drug signatures. (Default: LINCS L1000 Dec 2015)}

\item{refDrug}{Character vector specifying a reference drug signature to use from datasetInput_Dataset. (Default: JQ1)}

\item{n_bins}{Numeric value to filter out the lowest 'n' percentile of the gene consensus scores from the Reference Drug Signature. (Default value: 33)}

\item{datasetInput_Dis}{Specify a disease signature dataset. (Default: Glioblastoma TCGA)}

\item{...}{Optional arguments passed to read.table().}
}
\value{
a data frame
}
\description{
Reads input data files and calculates similarity score of each drug to reference, ratio of genes discordant to disease signature.
Returns a data frame of similarity scores.
}
\concept{Drugs_SigsR functions}
